explicit constructors for ptrT
git-svn-id: http://svn.miranda-ng.org/main/trunk@6174 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Skype/src/skype_transfers.cpp
@@ -131,7 +131,7 @@ void CSkypeProto::OnFileEvent(const ConversationRef &conversation, const Message
 	FileTransferParam ftp;
 	ftp.pfts.flags = PFTS_RECEIVING | PFTS_UNICODE;
 	ftp.pfts.hContact = this->AddContact(author, true);
-	ftp.pfts.totalFiles = transfers.size();
+	ftp.pfts.totalFiles = (int)transfers.size();
 	ftp.pfts.ptszFiles = new wchar_t*[transfers.size() + 1];
 
 	int nifc = 0;
